refactor(st-09065): modularize langgraph state helpers
