+++ planning/checklists/epic-09-story-tasks.md
@@ -2733,23 +2733,51 @@ Implementation notes:
 **Branch:** `refactor/st-09065-langgraph-state-modularization`
 
 ### Checklist
-- [ ] Create branch `refactor/st-09065-langgraph-state-modularization`
+- [x] Create branch `refactor/st-09065-langgraph-state-modularization`
 - [ ] Create draft PR with story ID in title
-- [ ] Define test strategy before implementation: cover runtime modularization and test-file modularization; first failing test should prove state-helper behavior remains stable while the oversized runtime and test files are split
-- [ ] Write or update the failing automated test before production changes when practical; if not practical, record why before implementation
-- [ ] Reduce `packages/core/src/langgraph/state.ts` below the 300 line planning cutoff by extracting focused internal modules for annotation helpers, reducer helpers, schema/state utilities, and shared public types behind a stable facade
-- [ ] Keep extracted production modules below the 300 line planning cutoff as well; do not satisfy the story by only shrinking the public facade and moving the bulk into a new oversized helper unless an explicit exception is documented in the story notes
-- [ ] Split state-helper coverage into focused test modules so annotation, reducer, and schema behavior no longer depends on a single oversized test surface
-- [ ] Preserve existing state-helper behavior, reducer semantics, and public imports
-- [ ] Add/update production code until focused tests pass, keeping test evidence in checklist notes and PR body
-- [ ] Record explicit-`any` warning deltas and file-size/responsibility improvements for touched state-helper modules in story docs
-- [ ] Add or update story documentation at `docs/st09065-langgraph-state-modularization.md` (or document why not required)
-- [ ] Assess residual test impact; add/update additional automated tests when needed, or document why no further tests are required
+- [x] Define test strategy before implementation: cover runtime modularization and test-file modularization; first failing test should prove state-helper behavior remains stable while the oversized runtime and test files are split
+- [x] Write or update the failing automated test before production changes when practical; if not practical, record why before implementation
+- [x] Reduce `packages/core/src/langgraph/state.ts` below the 300 line planning cutoff by extracting focused internal modules for annotation helpers, reducer helpers, schema/state utilities, and shared public types behind a stable facade
+- [x] Keep extracted production modules below the 300 line planning cutoff as well; do not satisfy the story by only shrinking the public facade and moving the bulk into a new oversized helper unless an explicit exception is documented in the story notes
+- [x] Split state-helper coverage into focused test modules so annotation, reducer, and schema behavior no longer depends on a single oversized test surface
+- [x] Preserve existing state-helper behavior, reducer semantics, and public imports
+- [x] Add/update production code until focused tests pass, keeping test evidence in checklist notes and PR body
+- [x] Record explicit-`any` warning deltas and file-size/responsibility improvements for touched state-helper modules in story docs
+- [x] Add or update story documentation at `docs/st09065-langgraph-state-modularization.md` (or document why not required)
+- [x] Assess residual test impact; add/update additional automated tests when needed, or document why no further tests are required
 - [ ] Run full test suite before finalizing the PR and record results
 - [ ] Run lint (`pnpm lint`) before finalizing the PR and record results
 - [ ] Commit completed checklist items as logical commits and push updates
 - [ ] Mark PR Ready only after all story tasks are complete
 - [ ] Wait for merge; do not merge directly from local branch
+
+Implementation notes:
+
+- Branch created from `main`: `refactor/st-09065-langgraph-state-modularization`
+- Test-first strategy:
+  - This story is behavior-preserving modularization, so a literal failing test for "the state helper file got smaller" would only assert repository structure rather than state-helper behavior.
+  - Practical substitute: split `packages/core/tests/langgraph/state.test.ts` into focused suites first, keep the public `state.test.ts` entrypoint stable, run that unchanged entrypoint before the production refactor, then modularize `packages/core/src/langgraph/state.ts` behind the same public facade.
+- Focused validation before production split:
+  - `pnpm test --run packages/core/tests/langgraph/state.test.ts`
+  - `1` file passed, `22` tests passed
+- Runtime file-size result:
+  - `packages/core/src/langgraph/state.ts`: `361 -> 13` lines
+  - extracted modules: `129`, `50`, `58`, `28`, and `45` lines
+- Test modularization result:
+  - `packages/core/tests/langgraph/state.test.ts`: `507 -> 4` lines
+  - focused suites: `136`, `93`, `83`, and `163` lines
+- Focused validation after production split:
+  - `pnpm test --run packages/core/tests/langgraph/state.test.ts`
+  - `1` file passed, `22` tests passed
+- Compatibility validation:
+  - `pnpm --filter @agentforge/core typecheck`
+- Explicit-`any` baseline check:
+  - `pnpm lint:explicit-any:baseline`
+  - `84/289` warnings overall; `23/119` in `core`; no regression from this story
+- Residual test impact:
+  - Additional behavior-specific tests were not required beyond the suite split because the story preserved the public state-helper facade; the new suites keep annotation, validation, merge, and default-handling workflow coverage isolated behind the same `state.test.ts` entrypoint.
+- CI impact:
+  - No CI change required; the story keeps the existing public entrypoints and validation commands intact.
 
 ---
